Integrate useSignOut into useAuth hook
