+++ client/hooks/useAuth.ts
@@ -3,15 +3,19 @@
  * @returns {object} - An object containing:
  *   - `isAuthenticated`: Boolean indicating whether the user is currently authenticated.
  *   - `isLoading`: Boolean indicating if the authentication status check is in progress.
+ *   - `signOut`: A  mutationfunction to sign out the user.
  */
 
 import { auth } from '@/firebase/firebaseConfig';
 import { useState, useEffect } from 'react';
-import { onAuthStateChanged } from 'firebase/auth';
+import { onAuthStateChanged, signOut } from 'firebase/auth';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+import type { FirebaseError } from 'firebase/app';
 
 const useAuth = () => {
   const [isAuthenticated, setIsAuthenticated] = useState(false);
   const [isLoading, setIsLoading] = useState(true);
+  const queryClient = useQueryClient();
 
   useEffect(() => {
     const unsubscribe = onAuthStateChanged(auth, (user) => {
@@ -26,7 +30,17 @@ const useAuth = () => {
     return () => unsubscribe();
   }, []);
 
-  return { isAuthenticated, isLoading };
+  const signOutMutation = useMutation<void, FirebaseError, void>({
+    mutationFn: () => signOut(auth),
+    onSuccess: () => {
+      queryClient.removeQueries({ queryKey: ['userProfile'] });
+    },
+    onError: (error: FirebaseError) => {
+      console.error('Error signing out:', error);
+    },
+  });
+
+  return { isAuthenticated, isLoading, signOut: signOutMutation.mutate };
 };
 
 export default useAuth;
